requirements.txt
opencv-contrib-python
numpy
Pillow

pytest